+++ cupidone/controllers/__init__.py
@@ -3,6 +3,7 @@ from .add import *
 from .build import *
 from .migrate import *
 from .dump import *
+from .site import *
 from .version import *
 from .help import *
 from .composite import *