+++ cupidone/models/project.py
@@ -2,7 +2,7 @@ from typing import List
 
 from cupidone.common import *
 from cupidone.common import dt_to_iso
-from cupidone.dc import CompiledCard, CompiledChecklistItem
+from cupidone.dc import CompiledCard, CompiledChecklistItem, WebData
 from cupidone.views.base import AbstractView, ListView
 from cupidone.views.message import ErrorView, InfoView
 
@@ -197,5 +197,16 @@ class ProjectModel(AbstractModel):
         cards = self._list_cards()
         return ListView(list=cards)
 
+    def build_site(self):
+        build = dt_to_iso(self.tm.get_datetime())
+        title = self.fm.get_project_dir()
+        cards = self._list_cards()
+        data = WebData(
+            build=build,
+            title=title,
+            cards=cards
+        )
+        self.fm.build_site(data)
+        return InfoView("done")
 
 __all__ = ["ProjectModel"]